+++ docs/changelog.rst
@@ -15,6 +15,9 @@ v2.3.0 (Not yet released)
   for non-MSVC compilers).
   `#934 <https://github.com/pybind/pybind11/pull/934>`_.
 
+* Fixed casting of polymorphic classes which inherit from non-polymorphic bases.
+  `#1084 <https://github.com/pybind/pybind11/pull/1084>`_.
+
 v2.2.1 (September 14, 2017)
 -----------------------------------------------------